apps/processor/processor-atomizers/pyproject.toml
[project]
name = "processor-atomizers"
version = "0.1.0"
description = "Unified Processor Atomizers – consolidates PPTX, Excel, PDF, CSV, AI, and Cleanup atomizers into a single service."
requires-python = ">=3.12"
dependencies = [
    # FastAPI + ASGI
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    # gRPC
    "grpcio>=1.62",
    "grpcio-tools>=1.62",
    "grpcio-health-checking>=1.62",
    # Pydantic
    "pydantic>=2.0",
    "pydantic-settings>=2.1",
    # HTTP client
    "httpx>=0.27",
    # Document processing – PPTX
    "python-pptx>=1.0.2",
    "Pillow>=10.0",
    # Document processing – Excel
    "openpyxl>=3.1.2",
    # Document processing – PDF / OCR
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    # Document processing – CSV
    "pandas>=2.0.0",
    "chardet>=5.0.0",
    # AI integration (LiteLLM gateway)
    "pyyaml>=6.0",
    "tiktoken>=0.6",
    # Azure Blob Storage
    "azure-storage-blob>=12.19.0",
    # Scheduling (cleanup worker)
    "APScheduler>=3.10",
    # OpenTelemetry observability
    "opentelemetry-api>=1.23",
    "opentelemetry-sdk>=1.23",
    "opentelemetry-exporter-otlp>=1.23",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "opentelemetry-instrumentation-grpc>=0.44b0",
    # Dapr
    "dapr>=1.13",
    "dapr-ext-grpc>=1.13",
    "dapr-ext-fastapi>=1.13",
    # Structured logging
    "python-json-logger>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120
